rename sm_gpa to gpa
