+++ packages/chain/statemanager/gpa/block_fetchers.go
@@ -1,10 +1,10 @@
-package sm_gpa
+package gpa
 
 import (
 	"time"
 
 	"github.com/iotaledger/hive.go/ds/shrinkingmap"
-	"github.com/iotaledger/wasp/packages/chain/statemanager/sm_gpa/sm_gpa_utils"
+	"github.com/iotaledger/wasp/packages/chain/statemanager/gpa/sm_gpa_utils"
 	"github.com/iotaledger/wasp/packages/state"
 )
