+++ packages/chain/statemanager/gpa/block_request_callback.go
@@ -1,4 +1,4 @@
-package sm_gpa
+package gpa
 
 type (
 	brcIsValidFun          func() bool
