+++ packages/chain/statemanager/gpa/sm_gpa_utils/block_wal_empty.go
