+++ packages/chain/statemanager/gpa/sm_gpa_utils/interface.go
