+++ packages/chain/statemanager/gpa/sm_gpa_utils/read_only_store.go
