+++ packages/chain/statemanager/gpa/sm_inputs/chain_fetch_state_diff.go
