+++ packages/chain/statemanager/gpa/sm_inputs/consensus_block_produced.go
