+++ packages/chain/statemanager/gpa/sm_inputs/consensus_state_proposal.go
