+++ packages/chain/statemanager/gpa/sm_messages/get_block_message.go
