+++ packages/chain/statemanager/gpa/sm_messages/interface.go
