+++ packages/chain/statemanager/gpa/state_manager_parameters.go
@@ -1,7 +1,7 @@
 // Copyright 2020 IOTA Stiftung
 // SPDX-License-Identifier: Apache-2.0
 
-package sm_gpa
+package gpa
 
 import (
 	"time"
